# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ClearGlass Fraud Detection Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  fraud: 'fraud',
  legitimate: 'legitimate',
  review: 'review',
} as const;

export type TransactionReviewStatus = typeof TransactionReviewStatus[keyof typeof TransactionReviewStatus];


export const TransactionReviewStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cleared: 'cleared',
} as const;

export interface Transaction {
  id: number;
  amount: number;
  category: string;
  merchantName: string;
  cardLast4: string;
  timestamp: string;
  /** 0.0 to 1.0 fraud probability */
  fraudScore: number;
  status: TransactionStatus;
  reviewStatus: TransactionReviewStatus;
  country?: string;
  /** @nullable */
  ipAddress?: string | null;
}

export type XAIExplanationMethod = typeof XAIExplanationMethod[keyof typeof XAIExplanationMethod];


export const XAIExplanationMethod = {
  SHAP: 'SHAP',
  LIME: 'LIME',
} as const;

export type FeatureContributionDirection = typeof FeatureContributionDirection[keyof typeof FeatureContributionDirection];


export const FeatureContributionDirection = {
  positive: 'positive',
  negative: 'negative',
} as const;

export interface FeatureContribution {
  feature: string;
  value: string;
  contribution: number;
  direction: FeatureContributionDirection;
}

export interface XAIExplanation {
  method: XAIExplanationMethod;
  summary: string;
  features: FeatureContribution[];
}

export interface TransactionDetail {
  transaction: Transaction;
  explanation: XAIExplanation;
}

export interface TransactionList {
  data: Transaction[];
  total: number;
  page: number;
  limit: number;
}

export interface TransactionInput {
  amount: number;
  category: string;
  merchantName: string;
  cardLast4: string;
  country: string;
  ipAddress?: string;
}

export type ReviewInputReviewStatus = typeof ReviewInputReviewStatus[keyof typeof ReviewInputReviewStatus];


export const ReviewInputReviewStatus = {
  confirmed: 'confirmed',
  cleared: 'cleared',
} as const;

export interface ReviewInput {
  reviewStatus: ReviewInputReviewStatus;
}

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Alert {
  id: number;
  title: string;
  description: string;
  severity: AlertSeverity;
  /** @nullable */
  transactionId: number | null;
  createdAt: string;
  resolved: boolean;
  /** @nullable */
  resolvedAt?: string | null;
}

export type AlertInputSeverity = typeof AlertInputSeverity[keyof typeof AlertInputSeverity];


export const AlertInputSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface AlertInput {
  title: string;
  description: string;
  severity: AlertInputSeverity;
  transactionId?: number;
}

export type ModelType = typeof ModelType[keyof typeof ModelType];


export const ModelType = {
  XGBoost: 'XGBoost',
  LightGBM: 'LightGBM',
  CatBoost: 'CatBoost',
  IsolationForest: 'IsolationForest',
  Ensemble: 'Ensemble',
} as const;

export interface Model {
  id: number;
  name: string;
  type: ModelType;
  version: string;
  isActive: boolean;
  trainedAt: string;
  description?: string;
}

export type ModelMetricsConfusionMatrix = {
  tp?: number;
  fp?: number;
  tn?: number;
  fn?: number;
};

export interface ModelMetrics {
  modelId: number;
  accuracy: number;
  aucRoc: number;
  precision: number;
  recall: number;
  f1Score: number;
  prAuc: number;
  confusionMatrix?: ModelMetricsConfusionMatrix;
}

export interface DashboardSummary {
  totalTransactions: number;
  fraudDetected: number;
  totalAlerts: number;
  activeAlerts: number;
  fraudRate: number;
  avgFraudScore: number;
  /** Average hours to resolve */
  investigationTurnaround: number;
  totalAmountAtRisk: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  fraud_detected: 'fraud_detected',
  alert_created: 'alert_created',
  transaction_cleared: 'transaction_cleared',
  model_updated: 'model_updated',
  synthetic_completed: 'synthetic_completed',
} as const;

export type ActivityItemSeverity = typeof ActivityItemSeverity[keyof typeof ActivityItemSeverity];


export const ActivityItemSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
  info: 'info',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  message: string;
  timestamp: string;
  severity: ActivityItemSeverity;
  /** @nullable */
  transactionId?: number | null;
}

export interface TrendPoint {
  date: string;
  fraudCount: number;
  legitimateCount: number;
  totalAmount: number;
}

export interface CategoryBreakdown {
  category: string;
  fraudCount: number;
  legitimateCount: number;
  fraudRate: number;
}

export type SyntheticJobMethod = typeof SyntheticJobMethod[keyof typeof SyntheticJobMethod];


export const SyntheticJobMethod = {
  GAN: 'GAN',
  VAE: 'VAE',
  SMOTE: 'SMOTE',
  ADASYN: 'ADASYN',
  Faker: 'Faker',
} as const;

export type SyntheticJobStatus = typeof SyntheticJobStatus[keyof typeof SyntheticJobStatus];


export const SyntheticJobStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface SyntheticJob {
  id: number;
  method: SyntheticJobMethod;
  rowCount: number;
  fraudRate: number;
  status: SyntheticJobStatus;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type SyntheticJobInputMethod = typeof SyntheticJobInputMethod[keyof typeof SyntheticJobInputMethod];


export const SyntheticJobInputMethod = {
  GAN: 'GAN',
  VAE: 'VAE',
  SMOTE: 'SMOTE',
  ADASYN: 'ADASYN',
  Faker: 'Faker',
} as const;

export interface SyntheticJobInput {
  method: SyntheticJobInputMethod;
  /**
     * @minimum 100
     * @maximum 100000
     */
  rowCount: number;
  /**
     * @minimum 0.001
     * @maximum 0.5
     */
  fraudRate: number;
  notes?: string;
}

export type ListTransactionsParams = {
page?: number;
limit?: number;
status?: ListTransactionsStatus;
minAmount?: number;
maxAmount?: number;
};

export type ListTransactionsStatus = typeof ListTransactionsStatus[keyof typeof ListTransactionsStatus];


export const ListTransactionsStatus = {
  all: 'all',
  fraud: 'fraud',
  legitimate: 'legitimate',
  review: 'review',
} as const;

export type ListAlertsParams = {
severity?: ListAlertsSeverity;
resolved?: boolean;
};

export type ListAlertsSeverity = typeof ListAlertsSeverity[keyof typeof ListAlertsSeverity];


export const ListAlertsSeverity = {
  all: 'all',
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type GetDashboardActivityParams = {
limit?: number;
};