# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ClearGlass Fraud Detection Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemSeverity } from './activityItemSeverity';
import type { ActivityItemType } from './activityItemType';

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  message: string;
  timestamp: string;
  severity: ActivityItemSeverity;
  /** @nullable */
  transactionId?: number | null;
}