# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ClearGlass Fraud Detection Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalTransactions: number;
  fraudDetected: number;
  totalAlerts: number;
  activeAlerts: number;
  fraudRate: number;
  avgFraudScore: number;
  /** Average hours to resolve */
  investigationTurnaround: number;
  totalAmountAtRisk: number;
}