# lib/api-zod/src/generated/types/featureContribution.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ClearGlass Fraud Detection Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { FeatureContributionDirection } from './featureContributionDirection';

export interface FeatureContribution {
  feature: string;
  value: string;
  contribution: number;
  direction: FeatureContributionDirection;
}